feat: 一些 Bug 修复和脚本处理相关

+++ src/class/appmanager.h
@@ -38,7 +38,7 @@ public:
         QApplication::tr("WingCloudStudio");
     }
 
-    uint currentMSecsSinceEpoch();
+    quint64 currentMSecsSinceEpoch();
 
 public slots:
     void openFile(const QString &file, bool autoDetect = true);
